feat: Change the running mode to live stream.
