+++ googlecloud/ml-person-detector/mediapipe_detector.py
@@ -1,5 +1,9 @@
 # -*- coding: utf-8 -*-
 """ Person Detector
+Inspired by
+ 1. Google AI Edge https://ai.google.dev/edge/mediapipe/solutions/vision/object_detector
+ 2. Mediapipe Github project https://github.com/google-ai-edge/mediapipe-samples/blob/main/examples/object_detection/python/object_detector.ipynb
+
 
 # Copyright 2025 GoogleCloud.fr LLC. All Rights Reserved.
 
@@ -32,8 +36,8 @@ pip install -r requirements.txt
 Then download an off-the-shelf model. Check out the [MediaPipe documentation](https://ai.google.dev/edge/mediapipe/solutions/vision/object_detector#efficientdet-lite2_model) for more image classification models that you can use.
 
 ```bash
-wget  -O efficientdet.tflite -q https://storage.googleapis.com/mediapipe-models/object_detector/efficientdet_lite0/int8/1/efficientdet_lite0.tflite
-wget  -O efficientdet.tflite  https://storage.googleapis.com/mediapipe-models/object_detector/efficientdet_lite0/float32/latest/efficientdet_lite0.tflite
+wget  -O efficientdet.tflite https://storage.googleapis.com/mediapipe-models/object_detector/efficientdet_lite0/int8/1/efficientdet_lite0.tflite
+wget  -O efficientdet.tflite https://storage.googleapis.com/mediapipe-models/object_detector/efficientdet_lite0/float32/latest/efficientdet_lite0.tflite
 # this is a  more accurate model than the above
 wget  -O efficientdet_lite2.tflite  https://storage.googleapis.com/mediapipe-models/object_detector/efficientdet_lite2/float32/latest/efficientdet_lite2.tflite
 ```
@@ -42,6 +46,8 @@ wget  -O efficientdet_lite2.tflite  https://storage.googleapis.com/mediapipe-mod
 import os, sys
 import cv2
 import numpy as np
+import time
+from gcp_utils import gcs_upload_blob, gcs_path_to_http_url  # Import the functions
 
 # Parameters must be set for each environment variable
 parameters = ["RTSP_URL", "TFLITE_MODEL_PATH", "MAX_WORKERS", "OUTPUT_IMAGE_DIR", "SKIP_X_FRAMES",
@@ -158,33 +164,28 @@ if not is_debugger_attached():
 # Implement asynchronous processing for image detection
 import concurrent.futures
 import queue
-
-results_queue = queue.Queue()
-# Thread pool for asynchronous tasks of image object detection
-detector_executor = concurrent.futures.ThreadPoolExecutor(max_workers=MAX_WORKERS)  # Adjust per CPU core as needed
-# Thread pool for uploading files to Google cloud storage
 import math
 
-gcs_executor = concurrent.futures.ThreadPoolExecutor(max_workers=math.ceil(MAX_WORKERS / 4))
+results_queue = queue.Queue()
+# Thread pool for uploading files to Google cloud storage
+gcs_executor = concurrent.futures.ThreadPoolExecutor(max_workers=math.ceil(MAX_WORKERS))
+
 
 # STEP 2: Create an ObjectDetector object.
-base_options = python.BaseOptions(model_asset_path=TFLITE_MODEL_PATH)
-options = vision.ObjectDetectorOptions(base_options=base_options, score_threshold=OBJ_DETECT_CONFIDENCE_SCORE)
-detector = vision.ObjectDetector.create_from_options(options)
-
-
-def detect_and_save_img(image, frame, PERSON_CONFIDENCE_SCORE=0.5):
-    detection_result = detector.detect(image)
+def detector_callback(detection_result: vision.ObjectDetectorResult,
+                      output_image: mp.Image, timestamp_ms: int):
+    person_detect_confidence_score = PERSON_DETECT_CONFIDENCE_SCORE
+    detection_result.timestamp_ms = timestamp_ms
     person_detected = False  # Flag to track person detection
     for detection in detection_result.detections:
         category = detection.categories[0]
-        if category.category_name == 'person' and category.score > PERSON_CONFIDENCE_SCORE:
+        if category.category_name == 'person' and category.score > person_detect_confidence_score:
             person_detected = True
             break  # No need to check further if a person is found
 
     if person_detected:  # Save image only if a person is detected
         print(f"Person detected: {detection_result}")
-        annotated_frame = visualize(frame, detection_result)  # Only visualize if saving
+        annotated_frame = visualize(output_image.numpy_view().copy(), detection_result)  # Only visualize if saving
         formatted_datetime = datetime.datetime.now().strftime("%Y-%m-%d_%H-%M-%S.%f")
         filename = f"annotated_frame_{formatted_datetime}.png"
         if not cv2.imwrite(f"{OUTPUT_IMAGE_DIR}/{filename}", annotated_frame):  # check for successful write
@@ -199,71 +200,10 @@ def detect_and_save_img(image, frame, PERSON_CONFIDENCE_SCORE=0.5):
         results_queue.put((None, None))  # Signal that no person detected, helps prevent queue buildup
 
 
-# upload detected images to GCS buckets in a async thread
-from google.cloud import storage
-
-
-def gcs_upload_blob(bucket_name, source_file_name, destination_blob_name):
-    """
-    Uploads a file to the bucket.
-    :param bucket_name: Google cloud storage bucket name
-    :param source_file_name: /local/path/to/file.png
-    :param destination_blob_name: folder-path/storage-object-name.png
-    :return: gs://bucket_name/destination_blob_name
-    """
-    try:
-        storage_client = storage.Client()
-        bucket = storage_client.bucket(bucket_name)
-        blob = bucket.blob(destination_blob_name)
-
-        # Optional: set a generation-match precondition to avoid potential race conditions
-        # and data corruptions. The request to upload is aborted if the object's
-        # generation number does not match your precondition. For a destination
-        # object that does not yet exist, set the if_generation_match precondition to 0.
-        # If the destination object already exists in your bucket, set instead a
-        # generation-match precondition using its generation number.
-        generation_match_precondition = 0
-
-        blob.upload_from_filename(source_file_name, if_generation_match=generation_match_precondition)
-        gs_path = f"gs://{bucket_name}/{destination_blob_name}"
-
-        return gs_path
-
-    except Exception as e:
-        sys.stderr.write(
-            f"Google Cloud Storage upload failed for {source_file_name} to {destination_blob_name} in bucket {bucket_name}: {e}!\n")  # Log or handle
-        return None  # Or return a specific error value
-
-
-def gcs_path_to_http_url(gcs_path):
-    """Converts a Google Cloud Storage path to a publicly accessible HTTP URL.
-
-    Args:
-        gcs_path: The Google Cloud Storage path (e.g., gs://bucket/path/to/file.txt)
-
-    Returns:
-        A publicly accessible HTTP URL for the GCS object,
-        or None if the path is not valid.
-
-    Raises:
-        ValueError: If the GCS path is not formatted correctly.
-    """
-
-    # Check if the path starts with "gs://"
-    if not gcs_path.startswith("gs://"):
-        raise ValueError("Invalid GCS path. Must start with 'gs://'")
-
-    # Extract bucket and object names
-    parts = gcs_path.split("/", 3)
-    if len(parts) < 4:
-        raise ValueError("Invalid GCS path format")
-
-    bucket_name = parts[2]
-    object_name = "/".join(parts[3:])
-
-    # Publicly accessible objects don't require additional configuration
-    return f"https://storage.cloud.google.com/{bucket_name}/{object_name}"
-
+base_options = python.BaseOptions(model_asset_path=TFLITE_MODEL_PATH)
+options = vision.ObjectDetectorOptions(base_options=base_options, running_mode=vision.RunningMode.LIVE_STREAM,
+                                       score_threshold=OBJ_DETECT_CONFIDENCE_SCORE, result_callback=detector_callback)
+detector = vision.ObjectDetector.create_from_options(options)
 
 # STEP 3: Load the input video, Detect objects in the input image.
 
@@ -285,9 +225,7 @@ if __name__ == "__main__":
             if frame_count % (frames_to_skip + 1) == 0:  # Process only every (frames_to_skip + 1)th frame
                 image = mp.Image(image_format=mp.ImageFormat.SRGB, data=frame)
                 # Detect objects in the current frame.
-                # Submit detection and saving task to the executor
-                detector_executor.submit(detect_and_save_img, image, frame.copy(),
-                                         PERSON_DETECT_CONFIDENCE_SCORE)  # .copy() important to prevent data corruption
+                detector.detect_async(image, int(time.time_ns() / 1000000))
 
             # Process results from the queue (non-blocking)
             if is_debugger_attached():
@@ -295,7 +233,7 @@ if __name__ == "__main__":
                       f"image detection and uploading to Google cloud storage: {threading.active_count()}", end="\r")
             try:
                 detection_result, filename = results_queue.get_nowait()
-                if detection_result:  # This is where you print. You'll only get here if a person was detected
+                if detection_result:  # Execution gets here if a person was detected
                     # Submit GCS upload to the executor
                     formatted_datetime = datetime.datetime.now().strftime("%Y-%m-%d_%H-%M-%S.%f")
                     gcs_folder_date_hour = (f"{formatted_datetime.split('_')[0]}/"
